refactor: move branches into src/features/branches + add @/ alias
