+++ electron.vite.config.ts
@@ -1,6 +1,7 @@
 import { defineConfig } from 'electron-vite'
 import react from '@vitejs/plugin-react'
 import tailwindcss from '@tailwindcss/vite'
+import path from 'node:path'
 import pkg from './package.json'
 
 export default defineConfig({
@@ -19,6 +20,7 @@ export default defineConfig({
   renderer: {
     root: '.',
     plugins: [react(), tailwindcss()],
+    resolve: { alias: { '@': path.resolve(__dirname, 'src') } },
     build: { rollupOptions: { input: 'index.html' } },
     define: { __APP_VERSION__: JSON.stringify(pkg.version) },
   },
