+++ src/features/branches/exports.ts
@@ -1,0 +1,1 @@
+export { BranchList } from './components/BranchList/BranchList'
