+++ vite.config.ts
@@ -2,12 +2,14 @@
 import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
 import tailwindcss from '@tailwindcss/vite'
+import path from 'node:path'
 import pkg from './package.json'
 
 export default defineConfig({
   base: '/donna/',
   server: { port: 5173 },
   plugins: [react(), tailwindcss()],
+  resolve: { alias: { '@': path.resolve(__dirname, 'src') } },
   define: { __APP_VERSION__: JSON.stringify(pkg.version) },
   test: {
     environment: 'jsdom',
